refactor(staking): move drift banner into staking banner notifications folder
