+++ packages/staking/src/features/overview/StakingNotificationBanners/PortfolioDriftBanner.css.ts
@@ -1,5 +1,5 @@
 import { style } from '@vanilla-extract/css';
-import { theme } from '../theme';
+import { theme } from '../../theme';
 
 export const bannerBellIcon = style({
   color: theme.colors.$bannerBellIconColor,
