--- packages/staking/src/features/portfolio-drift/index.ts
@@ -1,2 +1,0 @@
-export { isPortfolioDrifted } from './isPortfolioDrifted';
-export { PortfolioDriftBanner } from './PortfolioDriftBanner';
