+++ packages/staking/src/features/store/delegationPortfolioStore/isPortfolioDrifted.ts
@@ -2,8 +2,10 @@
 import { Wallet } from '@lace/cardano';
 import BigNumber from 'bignumber.js';
 import sum from 'lodash/sum';
-import type { CurrentPortfolioStakePool } from '../store';
-import { PERCENTAGE_SCALE_MAX, sumPercentagesSanitized } from '../store';
+import type { CurrentPortfolioStakePool } from './stateMachine/types';
+import { PERCENTAGE_SCALE_MAX } from './constants';
+import { isPortfolioSavedOnChain } from './isPortfolioSavedOnChain';
+import { sumPercentagesSanitized } from './stateMachine';
 
 const PORTFOLIO_DRIFT_PERCENTAGE_THRESHOLD = 15;
 
@@ -22,19 +24,13 @@ const getPortfolioTotalPercentageDrift = (
   );
 };
 
-const isPortfolioHavingSavedPercentages = (
-  currentPortfolio: CurrentPortfolioStakePool[]
-): currentPortfolio is (CurrentPortfolioStakePool & { savedIntegerPercentage: number })[] =>
-  currentPortfolio.every(({ savedIntegerPercentage }) => savedIntegerPercentage);
-
-// TODO: move this file to store. It gets imported also outside the overview feature so the store seems better place.
 export const isPortfolioDrifted = (currentPortfolio: CurrentPortfolioStakePool[]) => {
   const onChainPercentageSum = sumPercentagesSanitized({
     items: currentPortfolio,
     key: 'onChainPercentage',
   });
   if (onChainPercentageSum !== PERCENTAGE_SCALE_MAX) return true;
-  if (!isPortfolioHavingSavedPercentages(currentPortfolio)) return false;
+  if (!isPortfolioSavedOnChain(currentPortfolio)) return false;
   const drift = getPortfolioTotalPercentageDrift(currentPortfolio);
   return drift >= PORTFOLIO_DRIFT_PERCENTAGE_THRESHOLD;
 };
