+++ packages/staking/src/features/store/delegationPortfolioStore/isPortfolioSavedOnChain.ts
@@ -1,0 +1,6 @@
+import type { CurrentPortfolioStakePool } from './stateMachine/types';
+
+export const isPortfolioSavedOnChain = (
+  currentPortfolio: CurrentPortfolioStakePool[]
+): currentPortfolio is (CurrentPortfolioStakePool & { savedIntegerPercentage: number })[] =>
+  currentPortfolio.every(({ savedIntegerPercentage }) => savedIntegerPercentage);
